requirements.txt
# Core ML and AI libraries
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# LangChain ecosystem
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1
langchain-groq>=0.1.0

# LLM providers
groq>=0.4.0

# PDF processing
pypdf>=4.0.0

# Web framework
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0

# Optional: Google AI (for future use)
# langchain-google-genai>=1.0.0

# Development/Security tools (for CI/CD)
# bandit>=1.7.0
# safety>=3.0.0